+++ src/HelperTypes.ts
@@ -8,6 +8,9 @@ export interface Dataset {
   /** A function returning the color of the stroke given an input opacity value. */
   color?: (opacity: number) => string;
 
+  /** A function returning array of the colors of the stroke given an input opacity value for each data value. */
+  colors?: Array<(opacity: number) => string>;
+
   /** The width of the stroke. Defaults to 2. */
   strokeWidth?: number;
 
@@ -21,6 +24,14 @@ export interface Dataset {
   ignoreValue?: number;
 
   withShadow?: boolean;
+  /** Unique key **/
+  key?: string | number;
+
+  /** Stroke Dash Array */
+  strokeDashArray?: number[];
+
+  /** Stroke Dash Offset */
+  strokeDashOffset?: number;
 }
 
 export interface ChartData {