+++ src/ProgressChart.tsx
@@ -10,15 +10,12 @@ import AbstractChart, {
 
 export type ProgressChartData =
   | Array<number>
-  | { labels?: Array<string>; data: Array<number> };
+  | { labels?: Array<string>; colors?: Array<string>; data: Array<number> };
 
 export interface ProgressChartProps extends AbstractChartProps {
   data: ProgressChartData;
   width: number;
   height: number;
-  accessor: string;
-  backgroundColor: string;
-  paddingLeft: string;
   center?: Array<number>;
   absolute?: boolean;
   hasLegend?: boolean;
@@ -27,6 +24,7 @@ export interface ProgressChartProps extends AbstractChartProps {
   hideLegend?: boolean;
   strokeWidth?: number;
   radius?: number;
+  withCustomBarColorFromData?: boolean;
 }
 
 type ProgressChartState = {};
@@ -94,6 +92,9 @@ class ProgressChart extends AbstractChart<
     const withLabel = (i: number) =>
       (data as any).labels && (data as any).labels[i];
 
+    const withColor = (i: number) =>
+      (data as any).colors && (data as any).colors[i];
+
     const legend = !hideLegend && (
       <>
         <G>
@@ -103,7 +104,11 @@ class ProgressChart extends AbstractChart<
                 key={Math.random()}
                 width="16px"
                 height="16px"
-                fill={this.props.chartConfig.color(0.2 * (i + 1), i)}
+                fill={
+                  this.props.withCustomBarColorFromData
+                    ? withColor(i)
+                    : this.props.chartConfig.color(0.2 * (i + 1), i)
+                }
                 rx={8}
                 ry={8}
                 x={this.props.width / 2.5 - 24}
@@ -170,7 +175,10 @@ class ProgressChart extends AbstractChart<
             ry={borderRadius}
             fill="url(#backgroundGradient)"
           />
-          <G x={this.props.width / (hideLegend ? 2 : 2.5)} y={this.props.height / 2}>
+          <G
+            x={this.props.width / (hideLegend ? 2 : 2.5)}
+            y={this.props.height / 2}
+          >
             <G>
               {pieBackgrounds.map((pie, i) => {
                 return (
@@ -192,10 +200,14 @@ class ProgressChart extends AbstractChart<
                     strokeLinejoin="round"
                     d={pie.curves[0].sector.path.print()}
                     strokeWidth={strokeWidth}
-                    stroke={this.props.chartConfig.color(
-                      (i / pies.length) * 0.5 + 0.5,
-                      i
-                    )}
+                    stroke={
+                      this.props.withCustomBarColorFromData
+                        ? withColor(i)
+                        : this.props.chartConfig.color(
+                            (i / pies.length) * 0.5 + 0.5,
+                            i
+                          )
+                    }
                   />
                 );
               })}