+++ src/line-chart/LineChart.tsx
@@ -174,6 +174,7 @@ export interface LineChartProps extends AbstractChartProps {
     x: number;
     y: number;
     index: number;
+    indexData: number;
   }) => React.ReactNode;
   /**
    * Rotation angle of the horizontal labels - default 0 (degrees).
@@ -334,7 +335,7 @@ class LineChart extends AbstractChart<LineChartProps, LineChartState> {
             fillOpacity={0}
             onPress={onPress}
           />,
-          renderDotContent({ x: cx, y: cy, index: i })
+          renderDotContent({ x: cx, y: cy, index: i, indexData: x })
         );
       });
     });
@@ -643,6 +644,8 @@ class LineChart extends AbstractChart<LineChartProps, LineChartState> {
           fill="none"
           stroke={this.getColor(dataset, 0.2)}
           strokeWidth={this.getStrokeWidth(dataset)}
+          strokeDasharray={dataset.strokeDashArray}
+          strokeDashoffset={dataset.strokeDashOffset}
         />
       );
     });
@@ -738,6 +741,8 @@ class LineChart extends AbstractChart<LineChartProps, LineChartState> {
           fill="none"
           stroke={this.getColor(dataset, 0.2)}
           strokeWidth={this.getStrokeWidth(dataset)}
+          strokeDasharray={dataset.strokeDashArray}
+          strokeDashoffset={dataset.strokeDashOffset}
         />
       );
     });